Program #2
This program generates the first n terms of the Fibonacci sequence and prints them out. You can change the value of n to generate more or fewer terms as desired.

The program uses an array to store the Fibonacci sequence. It initializes the first two values to 1, then uses a for loop to calculate each subsequent value as the sum of the two preceding values. Finally, it prints out the entire sequence using another for loop.